refactor: ProxyFactoryBean 리팩터링
ProxyFactoryBean 에서 특정 기술에 의존적인 코드를 제거하였다.
프록시를 생성할 때 DI 가 필요할 수 있으므로 BeanFactory 를 주입 받도록 하였다.
